url_shortener/requirements.txt
Django==5.1.6
djangorestframework==3.14.0
drf-yasg==1.21.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # Только если используете PostgreSQL